Make it clearer that sleep time is an offset
To make it clearer that the sleep function is setting an offset from
the current time, insert the word 'in' between the action selector
and the time picker. This raises the possibility of using 'at' in
the same location and allowing absolute times to be set, so add a
little bit of code to support that but make it invisible for now.