Show the elapsed time in mm:ss format, even for tracks longer than 60 min
The track length is always shown as mm:ss; showing the elapsed time in a
different format (hh:mm:ss) makes it harder to compare the two numbers.
In addition a hh:mm:ss string does not fit in the space available in the UI -
only the mm:ss part appears, which is _very_ confusing.